requirements.txt
imageio-ffmpeg=0.4.7
numpy=1.23.1
pandas=1.4.4
pynrrd=1.0.0
PyQt5=5.15.7
pyvista=0.36.1
pyvistaqt=0.9.0
tqdm=4.64.1
wget=3.2